requirements.txt
requests>=2.23.0
pillow>=8.3.0
ascii_magic>=1.5.5